fix: start page numbering from 1 instead of 0 (#40)
- Add WebConfig with PageableHandlerMethodArgumentResolver to handle page indexing from 1
- Update API response to return page numbers starting from 1
